+++ packages/editor/src/editor/plugin/pickPoint/PickPoint.ts
@@ -128,6 +128,34 @@ class PickPoint implements BasePluginType {
     });
   }
 
+  generatePointsFromPosition(positions: Array<{ x: number; y: number; z: number }>) {
+    if (positions.length === 0) {
+      return;
+    }
+
+    const pickPointService = this.#PickPointService;
+
+    positions.forEach(position => {
+      const sphere = pickPointService.makeSphere(true);
+      const mapPointLabel = pickPointService.makeMeasurementLabel('mapPoint', sphere);
+
+      sphere.add(mapPointLabel);
+
+      sphere.position.setX(position.x);
+      sphere.position.setY(position.y);
+      sphere.position.setZ(position.z);
+
+      this.#Editor.scene.add(sphere);
+      this.#generatedPoints.push(sphere);
+
+      this.#Editor.trigger('point_initialize', {
+        type: 'point_initialize',
+        target: sphere,
+        id: sphere.id,
+      });
+    });
+  }
+
   #mousedownCallback = (event: MouseEvent) => {
     if (this.#isTransformDragging) {
       return;