+++ packages/editor/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.2](https://github.com/maxverse-dev/maxverse-web-sdk/compare/@maxverse/editor-web-sdk@0.0.1...@maxverse/editor-web-sdk@0.0.2) (2023-10-26)
+
+
+### Features
+
+* **plugin/PickPoint:** generate sphere material by using external position data ([cdc3a1c](https://github.com/maxverse-dev/maxverse-web-sdk/commit/cdc3a1cca20f4c2be7e0f8c10c00f421f74bc544))
+* **type/events:** add point initialize custom event ([8244f04](https://github.com/maxverse-dev/maxverse-web-sdk/commit/8244f046b7192c0c38324794790beb26d6cbb87b))
+
+
+
+
+
 ## 0.0.1 (2023-10-25)
 
 **Note:** Version bump only for package @maxverse/editor-web-sdk